Updated about.html
Improved HTML structure and syntax:
Adjusted meta tag (removed incorrect type attributes).
The <script> tag moved to the bottom of the <head> for better performance.
Simplified unnecessary nesting of elements in line 18.
Simplified the structure of download links (<a> tags) for resume downloads in lines: 186 and 189.
The function includeHtmlDocs() removed and was replaced by function includeRepeatedComponents();